Programmas pilnīga pārvietošana pa divām klasēm un dažādām metodēm
